+++ src/sql/run.py
@@ -25,6 +25,8 @@ import logging
 import warnings
 from collections.abc import Iterable
 
+DEFAULT_DISPLAYLIMIT_VALUE = 10
+
 
 def unduplicate_field_names(field_names):
     """Append a number to duplicate field names to make them unique."""
@@ -91,7 +93,7 @@ def _nonbreaking_spaces(match_obj):
     Make spaces visible in HTML by replacing all `` `` with ``&nbsp;``
 
     Call with a ``re`` match object.  Retain group 1, replace group 2
-    with nonbreaking speaces.
+    with nonbreaking spaces.
     """
     spaces = "&nbsp;" * len(match_obj.group(2))
     return "%s%s" % (match_obj.group(1), spaces)
@@ -154,12 +156,17 @@ class ResultSet(ColumnGuesserMixin):
             # to create clickable links
             result = html.unescape(result)
             result = _cell_with_spaces_pattern.sub(_nonbreaking_spaces, result)
-            if self.config.displaylimit and len(self) > self.config.displaylimit:
+            if len(self) > self.pretty.row_count:
                 HTML = (
                     '%s\n<span style="font-style:italic;text-align:center;">'
                     "%d rows, truncated to displaylimit of %d</span>"
+                    "<br>"
+                    '<span style="font-style:italic;text-align:center;">'
+                    "If you want to see more, please visit "
+                    '<a href="https://jupysql.ploomber.io/en/latest/api/configuration.html#displaylimit">displaylimit</a>'  # noqa: E501
+                    " configuration</span>"
                 )
-                result = HTML % (result, len(self), self.config.displaylimit)
+                result = HTML % (result, len(self), self.pretty.row_count)
             return result
         else:
             return None
@@ -172,7 +179,8 @@ class ResultSet(ColumnGuesserMixin):
             yield result
 
     def __str__(self, *arg, **kwarg):
-        self.pretty.add_rows(self)
+        if self.pretty:
+            self.pretty.add_rows(self)
         return str(self.pretty or "")
 
     def __repr__(self) -> str:
@@ -247,7 +255,7 @@ class ResultSet(ColumnGuesserMixin):
                       from each other in pie labels
         title: Plot title, defaults to name of value column
 
-        Any additional keyword arguments will be passsed
+        Any additional keyword arguments will be passed
         through to ``matplotlib.pylab.pie``.
         """
         self.guess_pie_columns(xlabel_sep=key_word_sep)
@@ -275,7 +283,7 @@ class ResultSet(ColumnGuesserMixin):
         ----------
         title: Plot title, defaults to names of Y value columns
 
-        Any additional keyword arguments will be passsed
+        Any additional keyword arguments will be passed
         through to ``matplotlib.pylab.plot``.
         """
         import matplotlib.pylab as plt
@@ -316,7 +324,7 @@ class ResultSet(ColumnGuesserMixin):
         key_word_sep: string used to separate column values
                       from each other in labels
 
-        Any additional keyword arguments will be passsed
+        Any additional keyword arguments will be passed
         through to ``matplotlib.pylab.bar``.
         """
         import matplotlib.pylab as plt
@@ -562,7 +570,7 @@ def raw_run(conn, sql):
 class PrettyTable(prettytable.PrettyTable):
     def __init__(self, *args, **kwargs):
         self.row_count = 0
-        self.displaylimit = None
+        self.displaylimit = DEFAULT_DISPLAYLIMIT_VALUE
         return super(PrettyTable, self).__init__(*args, **kwargs)
 
     def add_rows(self, data):
@@ -571,7 +579,7 @@ class PrettyTable(prettytable.PrettyTable):
         self.clear_rows()
         self.displaylimit = data.config.displaylimit
         if self.displaylimit == 0:
-            self.displaylimit = None  # TODO: remove this to make 0 really 0
+            self.displaylimit = None
         if self.displaylimit in (None, 0):
             self.row_count = len(data)
         else: